update Env-vars clean up code
