+++ .github/workflows/production_network_test.yml
@@ -125,7 +125,7 @@ jobs:
       with:
         what-to-test: $TEST_SET
       env:
-        Target: 'ProductionNetwork'
+        UC4_TESTBASE_TARGET: 'PRODUCTION_NETWORK'
 
     # Combine reports
     - name: Scalatest multiproject HTML merger
