+++ src/main/scala/de/upb/cs/uc4/hyperledger/utilities/helper/TransactionHelper.scala
@@ -4,10 +4,11 @@ import java.nio.charset.StandardCharsets
 
 import com.google.gson.Gson
 import com.google.protobuf.ByteString
+import de.upb.cs.uc4.hyperledger.connections.traits.ConnectionApprovalsTrait
 import org.hyperledger.fabric.gateway.impl.{ ContractImpl, TransactionImpl }
 import org.hyperledger.fabric.protos.common.Common
 import org.hyperledger.fabric.protos.peer.{ Chaincode, ProposalPackage }
-import org.hyperledger.fabric.protos.peer.ProposalPackage.{ ChaincodeProposalPayload, Proposal }
+import org.hyperledger.fabric.protos.peer.ProposalPackage.{ ChaincodeProposalPayload, Proposal, SignedProposal }
 import org.hyperledger.fabric.sdk.TransactionProposalRequest
 import org.hyperledger.fabric.sdk.transaction.TransactionContext
 
@@ -78,6 +79,21 @@ protected[hyperledger] object TransactionHelper {
     val chaincodeInput = invocationSpec.getChaincodeSpec.getInput
     val args: Array[ByteString] = chaincodeInput.getArgsList.asScala.toArray
     args.map[String]((b: ByteString) => new String(b.toByteArray, StandardCharsets.UTF_8)).toList
+  }
+
+  def createSignedProposal(approvalConnection: ConnectionApprovalsTrait, proposal: ProposalPackage.Proposal, signature: ByteString): (TransactionImpl, TransactionContext, SignedProposal) = {
+    val transactionId: String = TransactionHelper.getTransactionIdFromProposal(proposal)
+    val transactionName: String = TransactionHelper.getTransactionNameFromProposal(proposal)
+    val params: Seq[String] = TransactionHelper.getTransactionParamsFromProposal(proposal)
+
+    val signedProposalBuilder: SignedProposal.Builder = SignedProposal.newBuilder
+      .setProposalBytes(proposal.toByteString)
+      .setSignature(signature)
+    val signedProposal: SignedProposal = signedProposalBuilder.build
+
+    val (transaction, context, request) = TransactionHelper.createTransactionInfo(approvalConnection.contract, transactionName, params.toArray, Some(transactionId))
+
+    (transaction, context, signedProposal)
   }
 
 }
